adding convolve function
